# swagger.yaml
swagger: '2.0'
info:
  title: Custom URL Shortener API
  description: |
    A scalable URL Shortener API with advanced analytics, Google Sign-In authentication, and rate limiting.
  version: 1.0.0
servers:
  - url: http://localhost:9000
...
securityDefinitions:
  accessToken:
    type: apiKey
    in: header
    name: Authorization
    description: "Use the 'accessToken' in the header for authorization. Example: 'Authorization: Bearer <access_token>'"
paths:
...
  /api/analytics/overall:
    get:
      tags:
        - Analytics
      summary: Retrieve overall analytics for all short URLs
      description: Retrieve overall analytics for all short URLs created by the authenticated user, providing a comprehensive view of their link performance.
      operationId: getOverallAnalytics
      security:
        - accessToken: []  # Assuming Bearer token authentication for the user
      responses:
        '200':
          description: Successfully retrieved analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalUrls:
                    type: integer
                    description: Total number of short URLs created by the user.
                  totalClicks:
                    type: integer
                    description: Total number of clicks across all URLs created by the user.
                  uniqueClicks:
                    type: integer
                    description: Total number of unique users who accessed any of the user's short URLs.
                  clicksByDate:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                          description: Date of the clicks
                        totalClicks:
                          type: integer
                          description: Total number of clicks on that day
                  osType:
                    type: array
                    items:
                      type: object
                      properties:
                        osName:
                          type: string
                          description: Name of the operating system (e.g., Windows, macOS, Linux, iOS, Android)
                        uniqueClicks:
                          type: integer
                          description: Number of unique clicks for that OS
                        uniqueUsers:
                          type: integer
                          description: Number of unique users for that OS
                  deviceType:
                    type: array
                    items:
                      type: object
                      properties:
                        deviceName:
                          type: string
                          description: Name of the device used (e.g., mobile, desktop)
                        uniqueClicks:
                          type: integer
                          description: Number of unique clicks for that device type
                        uniqueUsers:
                          type: integer
                          description: Number of unique users for that device type
        '401':
          description: Unauthorized - Missing or invalid authentication token
        '500':
          description: Internal Server Error - Something went wrong on the server
      tags:
        - Analytics